requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
braceexpand==0.1.6
chardet==4.0.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@309c70222eec33acacb8c21bbb0fbe9bc57de661
discord-flags==2.1.1
discord.py==1.7.1
humanize==3.5.0
idna==3.1
import-expression==1.1.4
jishaku==1.20.0.220
multidict==5.1.0
Pillow==8.2.0
prettytable==2.1.0
psutil==5.8.0
python-dotenv==0.17.1
six==1.15.0
typing-extensions==3.10.0.0
wcwidth==0.2.5
yarl==1.6.3
